refactor: add separate export for edge runtimes
Split into local/remote, where "local" contains all the Node-specific filesystem code, whereas the "remote" part is platform independent.

Create new package export that targets edge runtimes by using the "worker" export condition.
