+++ packages/@tinacms/graphql/package.json
@@ -8,15 +8,26 @@
     "package.json",
     "dist"
   ],
-  "exports": {
-    "import": "./dist/index.mjs",
-    "require": "./dist/index.js"
+  "exports": {    
+    ".": {
+      "worker": "./dist/index-edge.mjs",
+      "import": "./dist/index.mjs",
+      "require": "./dist/index.js"
+    },
+    "./dist/index-edge": {
+      "import": "./dist/index-edge.mjs"
+    }
   },
   "license": "SEE LICENSE IN LICENSE",
   "buildConfig": {
     "entryPoints": [
       {
         "name": "src/index.ts",
+        "target": "node",
+        "bundle": []
+      },
+      {
+        "name": "src/index-edge.ts",
         "target": "node",
         "bundle": []
       }
