+++ packages/@tinacms/graphql/src/index-edge.ts
@@ -1,0 +1,2 @@
+export * from './database/remote'
+export { resolve } from './resolve'
